refactor user component
